Added es6 modules with Date object library
