+++ index.html
@@ -17,7 +17,7 @@
                 <li><a href="#contact" id="nav-contact">Contact</a></li>
             </ul>
         </nav>
-        <time datetime="2023-06-0T09:30">June 1st 2023, 09:30 am</time>
+        <time id="time"></time>
   </header>
 
     <main class="parent">
@@ -45,6 +45,6 @@
     <footer>
         <p class="copy-right">Copy Rights reserved to <br> Mohammed and Nico </p>
     </footer>
-        <script src="moduls/index.js"></script>
+        <script type="module" src="index.js"></script>
 </body>
 </html>
